Simplify ProtocolPluginFacade in Java and C#
