+++ cpp/src/Ice/ProtocolPluginFacade.cpp
@@ -11,20 +11,6 @@ using namespace std;
 using namespace Ice;
 using namespace IceInternal;
 
-IceInternal::ProtocolPluginFacade::~ProtocolPluginFacade() = default; // Out of line to avoid weak vtable
-
-ProtocolPluginFacadePtr
-IceInternal::getProtocolPluginFacade(const CommunicatorPtr& communicator)
-{
-    return make_shared<ProtocolPluginFacade>(communicator);
-}
-
-CommunicatorPtr
-IceInternal::ProtocolPluginFacade::getCommunicator() const
-{
-    return _communicator;
-}
-
 void
 IceInternal::ProtocolPluginFacade::addEndpointFactory(const EndpointFactoryPtr& factory) const
 {
@@ -38,7 +24,6 @@ IceInternal::ProtocolPluginFacade::getEndpointFactory(int16_t type) const
 }
 
 IceInternal::ProtocolPluginFacade::ProtocolPluginFacade(const CommunicatorPtr& communicator)
-    : _instance(getInstance(communicator)),
-      _communicator(communicator)
+    : _instance(getInstance(communicator))
 {
 }
