--- cpp/src/Ice/ProtocolPluginFacadeF.h
@@ -1,14 +1,0 @@
-// Copyright (c) ZeroC, Inc.
-
-#ifndef ICE_PROTOCOL_PLUGIN_FACADE_F_H
-#define ICE_PROTOCOL_PLUGIN_FACADE_F_H
-
-#include <memory>
-
-namespace IceInternal
-{
-    class ProtocolPluginFacade;
-    using ProtocolPluginFacadePtr = std::shared_ptr<ProtocolPluginFacade>;
-}
-
-#endif
