+++ csharp/src/Ice/Internal/Util.cs
@@ -6,9 +6,6 @@ public sealed class Util
 {
     public static Instance getInstance(Ice.Communicator communicator) => communicator.instance;
 
-    public static ProtocolPluginFacade getProtocolPluginFacade(Ice.Communicator communicator) =>
-        new ProtocolPluginFacadeI(communicator);
-
     public static ThreadPriority stringToThreadPriority(string s)
     {
         if (string.IsNullOrEmpty(s))
