+++ csharp/test/Ice/background/PluginI.cs
@@ -6,11 +6,11 @@ internal class PluginI : Ice.Plugin
 
     public void initialize()
     {
-        Ice.Internal.ProtocolPluginFacade facade = Ice.Internal.Util.getProtocolPluginFacade(_communicator);
+        var facade = new Ice.Internal.ProtocolPluginFacade(_communicator);
         for (short s = 0; s < 100; ++s)
         {
             Ice.Internal.EndpointFactory factory = facade.getEndpointFactory(s);
-            if (factory != null)
+            if (factory is not null)
             {
                 facade.addEndpointFactory(new EndpointFactory(factory));
             }
